requirements.txt
django
psycopg2-binary
djangorestframework
pillow